requirements.txt
flask
flask-cors
sentence-transformers
faiss-cpu==1.9.0.post1
PyPDF2
python-dotenv
requests
numpy
torch
